+++ src/main/groovy/projects/dktk/v2/operation.groovy
@@ -1,6 +1,7 @@
 package projects.dktk.v2
 
 import ca.uhn.fhir.model.api.TemporalPrecisionEnum
+import de.kairos.fhir.centraxx.metamodel.SurgeryComponent
 import org.hl7.fhir.r4.model.Procedure
 
 import static de.kairos.fhir.centraxx.metamodel.MultilingualEntry.LANG
@@ -27,6 +28,20 @@ procedure {
       system = "http://dktk.dkfz.de/fhir/onco/core/CodeSystem/SYSTTherapieartCS"
       code = "OP"
       display = "Operation"
+    }
+  }
+
+  code {
+    context.source[surgery().surgeryComponents()]?.each { final surgeryComponent ->
+      if (surgeryComponent[SurgeryComponent.CODE]) {
+        coding {
+          code = surgeryComponent[SurgeryComponent.CODE] as String
+          system = "http://fhir.de/CodeSystem/bfarm/ops"
+          if (surgeryComponent[SurgeryComponent.DATE]) {
+            version = getYear(surgeryComponent[SurgeryComponent.DATE] as String)
+          }
+        }
+      }
     }
   }
 
@@ -88,5 +103,9 @@ procedure {
  */
 static String normalizeDate(final String dateTimeString) {
   return dateTimeString != null ? dateTimeString.substring(0, 19) : null
+}
+
+static String getYear(final String dateTimeString) {
+  return dateTimeString != null ? dateTimeString.substring(0, 4) : null
 }